pyproject.toml
[tool.poetry]
name = "monitorizer"
version = "3.0.0"
description = ""
authors = ["Ahmed Ezzat <ahmed.ezzat119@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
celery = {extras = ["redis"], version = "^5.3.6"}
django-unfold = "^0.21.1"
django-celery-beat = "^2.6.0"
django-import-export = "^3.3.7"
requests = "^2.31.0"
markdown = "^3.6"
gunicorn = "^21.2.0"
psycopg = "^3.1.18"
django-debug-toolbar = "^4.3.0"
pytelegrambotapi = "^4.17.0"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
